fix dynamic values

+++ src/server/game/Entities/Object/Object.cpp
@@ -1164,7 +1164,7 @@ void Object::BuildDynamicValuesUpdate(uint8 updateType, ByteBuffer *data, Player
                 }
             }
 
-            fieldBuffer << uint8(arrayMask.GetBlockCount());
+            fieldBuffer << uint16(arrayMask.GetBlockCount());
             arrayMask.AppendToPacket(&fieldBuffer);
             fieldBuffer.append(buffer);
         }
